// enums/pages/navigationBarItem.ts
export enum NavigationBarItem {
    Admin = 'Admin',
    PIM = 'PIM',
    Leave = 'Leave',
    Time = 'Time',
    Recruitment = 'Recruitment',
    MyInfo = 'My Info',
    Performance = 'Performance',
    Dashboard = 'Dashboard',
    Directory = 'Directory',
    Maintenance = 'Maintenance',
    Claim = 'Claim',
    Buzz = 'Buzz'
}